# Use `Path2D`, if available, when rendering Type3-fonts (bug 810214)