@calixteman 2022-04-30
Would it be possible to build the paths in the worker ? would it be useful to do that if possible ?