@Snuffleupagus 2022-04-30
> Would it be possible to build the paths in the worker ?

Well, it doesn't actually seem possible to send a `Path2D`-object via `postMessage` according to some quick tests (it's not supported by the structured cloning algorithm as far as I can tell). Hence it seems like the answer is NO, and looking at all of the Type3-code in `canvas.js` it probably wouldn't have been exactly trivial anyway.

> would it be useful to do that if possible ?

I'm actually not so sure if would've helped much, had it worked, considering that:
 - Type3-fonts are, thankfully, not all that common in PDF documents. It seems that we're no longer collecting the relevant Telemetry-data, since [the most recent data I could find is a bit old](https://telemetry.mozilla.org/new-pipeline/dist.html#!cumulative=0&end_date=2021-07-12&include_spill=0&keys=__none__!__none__!__none__&max_channel_version=nightly%252F91&measure=PDF_VIEWER_FONT_TYPES_2&min_channel_version=nightly%252F75&processType=*&product=Firefox&sanitize=1&sort_by_value=0&sort_keys=submissions&start_date=2021-05-31&table=0&trim=1&use_submission_date=0).
 - Type3-fonts are already, thanks to the pre-existing compiling and caching on first use, rendered fairly quickly.
 - With this patch, the rendering time in real world test-cases is reduced even further.