@timvandermeij 2022-04-30
Out of curiosity, what makes `Path2D` so much faster here? AFAICT in the previous code we populated an array and looped over that at the end to perform the drawing instructions, while we now make a `Path2D` object where we do the same operations on so it seems, so it it purely the prevention of the array creation/loop that saves so much here or is `Path2D` doing something else more efficiently?