@Snuffleupagus 2022-04-30
> Is it purely the prevention of the array creation/loop that saves so much,

While I obviously don't *know*, this is likely more-or-less completely unrelated to the performance improvements.

> or is the fact that all drawing instructions are now being performed on the canvas object in one go instead of individually,

That'd rather be the most likely explanation, as far as I understand things.
In the old code we'd apply lots of moveTo/lineTo operators *individually* during rendering, whereas with Path2D all those instructions are applied at once instead (and thus allowing the browser to optimize things a fair bit). Also, everytime that we render one specific Type3-glyph we're now also using the same identical Path2D-object.